Correct data types for latitude & longitude fixes #44

+++ sql/CacheDB/CacheDB/dbo/Tables/CenterPoints.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[CenterPoints] (
     [Locationname] VARCHAR (20) NOT NULL,
-    [Latitude]     FLOAT (53)   NULL,
-    [Longitude]    FLOAT (53)   NULL,
+    [Latitude]     DECIMAL(8, 6)   NULL,
+    [Longitude]    DECIMAL(9, 6)   NULL,
     [latlong]      AS           ([geography]::Point([Latitude],[Longitude],(4326))) PERSISTED,
     CONSTRAINT [pk_centerpoints] PRIMARY KEY CLUSTERED ([Locationname] ASC)
 );
